split admin file into smaller files
