+++ src/bocken/models/__init__.py
@@ -1,6 +1,6 @@
 from .admin import Admin
 from .agreement import Agreement
-from .group import JournalEntryGroup
+from .journal_entry_group import JournalEntryGroup
 from .journal_entry import JournalEntry
 from .report import Report
